# services/auth/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateCredentialInputBody": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "description": "User email address",
            "type": "string"
          },
          "password": {
            "description": "Plaintext password; bcrypt hashed at cost 12",
            "minLength": 8,
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "type": "object"
      },
      "CreateCredentialOutputBody": {
        "additionalProperties": false,
        "properties": {
          "credential_id": {
            "description": "UUID of the created credential",
            "type": "string"
          }
        },
        "required": [
          "credential_id"
        ],
        "type": "object"
      },
...
      "ErrorModel": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "description": "A human-readable explanation specific to this occurrence of the problem.",
            "examples": [
              "Property foo is required but is missing."
            ],
            "type": "string"
          },
          "errors": {
            "description": "Optional list of individual error details",
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "instance": {
            "description": "A URI reference that identifies the specific occurrence of the problem.",
            "examples": [
              "https://example.com/error-log/abc123"
            ],
            "format": "uri",
            "type": "string"
          },
          "status": {
            "description": "HTTP status code",
            "examples": [
              400
            ],
            "format": "int64",
            "type": "integer"
          },
          "title": {
            "description": "A short, human-readable summary of the problem type. This value should not change between occurrences of the error.",
            "examples": [
              "Bad Request"
            ],
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "description": "A URI reference to human-readable documentation for the error.",
            "examples": [
              "https://example.com/errors/example"
            ],
            "format": "uri",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ValidateCredentialInputBody": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "description": "User email address",
            "type": "string"
          },
          "password": {
            "description": "Plaintext password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Veloci Auth",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/credentials/create": {
      "post": {
        "operationId": "create-credential",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCredentialInputBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCredentialOutputBody"
                }
              }
            },
            "description": "Created"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Create a new credential",
        "tags": [
          "credentials"
        ]
      }
    },
    "/credentials/validate": {
      "post": {
        "operationId": "validate-credential",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateCredentialInputBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateCredentialOutputBody"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Validate email/password credentials",
        "tags": [
          "credentials"
        ]
      }
    },
    "/credentials/{id}": {
      "delete": {
        "operationId": "delete-credential",
        "parameters": [
          {
            "description": "Credential UUID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "description": "Credential UUID",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Permanently remove a credential and cascade tokens",
        "tags": [
          "credentials"
        ]
      }
    },
    "/credentials/{id}/email": {
      "put": {
        "operationId": "update-credential-email",
        "parameters": [
          {
            "description": "Credential UUID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "description": "Credential UUID",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCredentialEmailInputBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Update email for a credential",
        "tags": [
          "credentials"
        ]
      }
    },
    "/credentials/{id}/password": {
      "put": {
        "operationId": "update-credential-password",
        "parameters": [
          {
            "description": "Credential UUID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "description": "Credential UUID",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCredentialPasswordInputBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Update password hash for a credential",
        "tags": [
          "credentials"
        ]
      }
    },
    "/invite": {
      "post": {
        "description": "The raw token is returned once and never stored. TTL from invite.ttl_hours in config (default 72h).",
        "operationId": "create-invite",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInviteInputBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateInviteOutputBody"
                }
              }
            },
            "description": "Created"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Create a one-time-use invite token",
        "tags": [
          "invite"
        ]
      }
    },
    "/invite/consume": {
      "post": {
        "description": "Sets accepted_at only if unconsumed. This is the commit point in the invite saga.",
        "operationId": "consume-invite",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsumeInviteInputBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Atomically consume an invite token",
        "tags": [
          "invite"
        ]
      }
    },
    "/tokens/mint": {
      "post": {
        "operationId": "mint-token",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MintTokenInputBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPairOutputBody"
                }
              }
            },
            "description": "Created"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Mint an access+refresh token pair",
        "tags": [
          "tokens"
        ]
      }
    },
    "/tokens/refresh": {
      "post": {
        "description": "Rotates the refresh token within a 30-second grace window to handle concurrent requests.",
        "operationId": "refresh-token",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenInputBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPairOutputBody"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Exchange a refresh token for a new access+refresh pair",
        "tags": [
          "tokens"
        ]
      }
    },
    "/tokens/user/{credential_id}": {
      "delete": {
        "operationId": "revoke-user-tokens",
        "parameters": [
          {
            "description": "Credential UUID",
            "in": "path",
            "name": "credential_id",
            "required": true,
            "schema": {
              "description": "Credential UUID",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Revoke all sessions for a credential without deleting it",
        "tags": [
          "tokens"
        ]
      }
    },
    "/tokens/validate": {
      "post": {
        "description": "Auth detects token type by structure (JWT = two dots; OTU = no dots) and queries the appropriate table.",
        "operationId": "validate-token",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateTokenInputBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateTokenOutputBody"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Validate any token — JWT or OTU",
        "tags": [
          "tokens"
        ]
      }
    },
    "/tokens/{jti}": {
      "delete": {
        "operationId": "revoke-token",
        "parameters": [
          {
            "description": "Token JTI",
            "in": "path",
            "name": "jti",
            "required": true,
            "schema": {
              "description": "Token JTI",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Revoke a single token by JTI (idempotent)",
        "tags": [
          "tokens"
        ]
      }
    }
  }
}