# services/web/authclient/oas_router_gen.go
// Code generated by ogen, DO NOT EDIT.

package authclient

import (
	"net/http"
	"net/url"
	"strings"

	"github.com/ogen-go/ogen/uri"
)

var (
	rn3AllowedHeaders = map[string]string{
		"POST": "Content-Type",
	}
	rn16AllowedHeaders = map[string]string{
		"POST": "Content-Type",
	}
	rn13AllowedHeaders = map[string]string{
		"PUT": "Content-Type",
	}
	rn15AllowedHeaders = map[string]string{
		"PUT": "Content-Type",
	}
	rn4AllowedHeaders = map[string]string{
		"POST": "Content-Type",
	}
	rn1AllowedHeaders = map[string]string{
		"POST": "Content-Type",
	}
	rn7AllowedHeaders = map[string]string{
		"POST": "Content-Type",
	}
	rn9AllowedHeaders = map[string]string{
		"POST": "Content-Type",
	}
	rn17AllowedHeaders = map[string]string{
		"POST": "Content-Type",
	}
)

func (s *Server) cutPrefix(path string) (string, bool) {
	prefix := s.cfg.Prefix
	if prefix == "" {
		return path, true
	}
	if !strings.HasPrefix(path, prefix) {
		// Prefix doesn't match.
		return "", false
	}
	// Cut prefix from the path.
	return strings.TrimPrefix(path, prefix), true
}

// ServeHTTP serves http request as defined by OpenAPI v3 specification,
// calling handler that matches the path or returning not found error.
func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	elem := r.URL.Path
	elemIsEscaped := false
	if rawPath := r.URL.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
			elemIsEscaped = strings.ContainsRune(elem, '%')
		}
	}

	elem, ok := s.cutPrefix(elem)
	if !ok || len(elem) == 0 {
		s.notFound(w, r)
		return
	}
	args := [1]string{}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/"

			if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'c': // Prefix: "credentials/"

				if l := len("credentials/"); len(elem) >= l && elem[0:l] == "credentials/" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'c': // Prefix: "create"
					origElem := elem
					if l := len("create"); len(elem) >= l && elem[0:l] == "create" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "POST":
							s.handleCreateCredentialRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "POST",
								allowedHeaders: rn3AllowedHeaders,
								acceptPost:     "application/json",
								acceptPatch:    "",
							})
						}

						return
					}

					elem = origElem
				case 'v': // Prefix: "validate"
					origElem := elem
					if l := len("validate"); len(elem) >= l && elem[0:l] == "validate" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "POST":
							s.handleValidateCredentialRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "POST",
								allowedHeaders: rn16AllowedHeaders,
								acceptPost:     "application/json",
								acceptPatch:    "",
							})
						}

						return
					}

					elem = origElem
				}
				// Param: "id"
				// Match until "/"
				idx := strings.IndexByte(elem, '/')
				if idx < 0 {
					idx = len(elem)
				}
				args[0] = elem[:idx]
				elem = elem[idx:]

				if len(elem) == 0 {
					switch r.Method {
					case "DELETE":
						s.handleDeleteCredentialRequest([1]string{
							args[0],
						}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "DELETE",
							allowedHeaders: nil,
							acceptPost:     "",
							acceptPatch:    "",
						})
					}

					return
				}
				switch elem[0] {
				case '/': // Prefix: "/"

					if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						break
					}
					switch elem[0] {
					case 'e': // Prefix: "email"

						if l := len("email"); len(elem) >= l && elem[0:l] == "email" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "PUT":
								s.handleUpdateCredentialEmailRequest([1]string{
									args[0],
								}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, notAllowedParams{
									allowedMethods: "PUT",
									allowedHeaders: rn13AllowedHeaders,
									acceptPost:     "",
									acceptPatch:    "",
								})
							}

							return
						}

					case 'p': // Prefix: "password"

						if l := len("password"); len(elem) >= l && elem[0:l] == "password" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "PUT":
								s.handleUpdateCredentialPasswordRequest([1]string{
									args[0],
								}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, notAllowedParams{
									allowedMethods: "PUT",
									allowedHeaders: rn15AllowedHeaders,
									acceptPost:     "",
									acceptPatch:    "",
								})
							}

							return
						}

					}

				}

			case 'i': // Prefix: "invite"

				if l := len("invite"); len(elem) >= l && elem[0:l] == "invite" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch r.Method {
					case "POST":
						s.handleCreateInviteRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "POST",
							allowedHeaders: rn4AllowedHeaders,
							acceptPost:     "application/json",
							acceptPatch:    "",
						})
					}

					return
				}
				switch elem[0] {
				case '/': // Prefix: "/consume"

					if l := len("/consume"); len(elem) >= l && elem[0:l] == "/consume" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "POST":
							s.handleConsumeInviteRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "POST",
								allowedHeaders: rn1AllowedHeaders,
								acceptPost:     "application/json",
								acceptPatch:    "",
							})
						}

						return
					}

				}

			case 't': // Prefix: "tokens/"

				if l := len("tokens/"); len(elem) >= l && elem[0:l] == "tokens/" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'm': // Prefix: "mint"
					origElem := elem
					if l := len("mint"); len(elem) >= l && elem[0:l] == "mint" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "POST":
							s.handleMintTokenRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "POST",
								allowedHeaders: rn7AllowedHeaders,
								acceptPost:     "application/json",
								acceptPatch:    "",
							})
						}

						return
					}

					elem = origElem
				case 'r': // Prefix: "refresh"
					origElem := elem
					if l := len("refresh"); len(elem) >= l && elem[0:l] == "refresh" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "POST":
							s.handleRefreshTokenRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "POST",
								allowedHeaders: rn9AllowedHeaders,
								acceptPost:     "application/json",
								acceptPatch:    "",
							})
						}

						return
					}

					elem = origElem
				case 'u': // Prefix: "user/"
					origElem := elem
					if l := len("user/"); len(elem) >= l && elem[0:l] == "user/" {
						elem = elem[l:]
					} else {
						break
					}

					// Param: "credential_id"
					// Leaf parameter, slashes are prohibited
					idx := strings.IndexByte(elem, '/')
					if idx >= 0 {
						break
					}
					args[0] = elem
					elem = ""

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "DELETE":
							s.handleRevokeUserTokensRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "DELETE",
								allowedHeaders: nil,
								acceptPost:     "",
								acceptPatch:    "",
							})
						}

						return
					}

					elem = origElem
				case 'v': // Prefix: "validate"
					origElem := elem
					if l := len("validate"); len(elem) >= l && elem[0:l] == "validate" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "POST":
							s.handleValidateTokenRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "POST",
								allowedHeaders: rn17AllowedHeaders,
								acceptPost:     "application/json",
								acceptPatch:    "",
							})
						}

						return
					}

					elem = origElem
				}
				// Param: "jti"
				// Leaf parameter, slashes are prohibited
				idx := strings.IndexByte(elem, '/')
				if idx >= 0 {
					break
				}
				args[0] = elem
				elem = ""

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "DELETE":
						s.handleRevokeTokenRequest([1]string{
							args[0],
						}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "DELETE",
							allowedHeaders: nil,
							acceptPost:     "",
							acceptPatch:    "",
						})
					}

					return
				}

			}

		}
	}
	s.notFound(w, r)
}

// Route is route object.
type Route struct {
	name           string
	summary        string
	operationID    string
	operationGroup string
	pathPattern    string
	count          int
	args           [1]string
}

// Name returns ogen operation name.
//
// It is guaranteed to be unique and not empty.
func (r Route) Name() string {
	return r.name
}

// Summary returns OpenAPI summary.
func (r Route) Summary() string {
	return r.summary
}

// OperationID returns OpenAPI operationId.
func (r Route) OperationID() string {
	return r.operationID
}

// OperationGroup returns the x-ogen-operation-group value.
func (r Route) OperationGroup() string {
	return r.operationGroup
}

// PathPattern returns OpenAPI path.
func (r Route) PathPattern() string {
	return r.pathPattern
}

// Args returns parsed arguments.
func (r Route) Args() []string {
	return r.args[:r.count]
}

// FindRoute finds Route for given method and path.
//
// Note: this method does not unescape path or handle reserved characters in path properly. Use FindPath instead.
func (s *Server) FindRoute(method, path string) (Route, bool) {
	return s.FindPath(method, &url.URL{Path: path})
}

// FindPath finds Route for given method and URL.
func (s *Server) FindPath(method string, u *url.URL) (r Route, _ bool) {
	var (
		elem = u.Path
		args = r.args
	)
	if rawPath := u.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
		}
		defer func() {
			for i, arg := range r.args[:r.count] {
				if unescaped, err := url.PathUnescape(arg); err == nil {
					r.args[i] = unescaped
				}
			}
		}()
	}

	elem, ok := s.cutPrefix(elem)
	if !ok {
		return r, false
	}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/"

			if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'c': // Prefix: "credentials/"

				if l := len("credentials/"); len(elem) >= l && elem[0:l] == "credentials/" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'c': // Prefix: "create"
					origElem := elem
					if l := len("create"); len(elem) >= l && elem[0:l] == "create" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "POST":
							r.name = CreateCredentialOperation
							r.summary = "Create a new credential"
							r.operationID = "create-credential"
							r.operationGroup = ""
							r.pathPattern = "/credentials/create"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

					elem = origElem
				case 'v': // Prefix: "validate"
					origElem := elem
					if l := len("validate"); len(elem) >= l && elem[0:l] == "validate" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "POST":
							r.name = ValidateCredentialOperation
							r.summary = "Validate email/password credentials"
							r.operationID = "validate-credential"
							r.operationGroup = ""
							r.pathPattern = "/credentials/validate"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

					elem = origElem
				}
				// Param: "id"
				// Match until "/"
				idx := strings.IndexByte(elem, '/')
				if idx < 0 {
					idx = len(elem)
				}
				args[0] = elem[:idx]
				elem = elem[idx:]

				if len(elem) == 0 {
					switch method {
					case "DELETE":
						r.name = DeleteCredentialOperation
						r.summary = "Permanently remove a credential and cascade tokens"
						r.operationID = "delete-credential"
						r.operationGroup = ""
						r.pathPattern = "/credentials/{id}"
						r.args = args
						r.count = 1
						return r, true
					default:
						return
					}
				}
				switch elem[0] {
				case '/': // Prefix: "/"

					if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						break
					}
					switch elem[0] {
					case 'e': // Prefix: "email"

						if l := len("email"); len(elem) >= l && elem[0:l] == "email" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch method {
							case "PUT":
								r.name = UpdateCredentialEmailOperation
								r.summary = "Update email for a credential"
								r.operationID = "update-credential-email"
								r.operationGroup = ""
								r.pathPattern = "/credentials/{id}/email"
								r.args = args
								r.count = 1
								return r, true
							default:
								return
							}
						}

					case 'p': // Prefix: "password"

						if l := len("password"); len(elem) >= l && elem[0:l] == "password" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch method {
							case "PUT":
								r.name = UpdateCredentialPasswordOperation
								r.summary = "Update password hash for a credential"
								r.operationID = "update-credential-password"
								r.operationGroup = ""
								r.pathPattern = "/credentials/{id}/password"
								r.args = args
								r.count = 1
								return r, true
							default:
								return
							}
						}

					}

				}

			case 'i': // Prefix: "invite"

				if l := len("invite"); len(elem) >= l && elem[0:l] == "invite" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch method {
					case "POST":
						r.name = CreateInviteOperation
						r.summary = "Create a one-time-use invite token"
						r.operationID = "create-invite"
						r.operationGroup = ""
						r.pathPattern = "/invite"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}
				switch elem[0] {
				case '/': // Prefix: "/consume"

					if l := len("/consume"); len(elem) >= l && elem[0:l] == "/consume" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "POST":
							r.name = ConsumeInviteOperation
							r.summary = "Atomically consume an invite token"
							r.operationID = "consume-invite"
							r.operationGroup = ""
							r.pathPattern = "/invite/consume"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

				}

			case 't': // Prefix: "tokens/"

				if l := len("tokens/"); len(elem) >= l && elem[0:l] == "tokens/" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'm': // Prefix: "mint"
					origElem := elem
					if l := len("mint"); len(elem) >= l && elem[0:l] == "mint" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "POST":
							r.name = MintTokenOperation
							r.summary = "Mint an access+refresh token pair"
							r.operationID = "mint-token"
							r.operationGroup = ""
							r.pathPattern = "/tokens/mint"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

					elem = origElem
				case 'r': // Prefix: "refresh"
					origElem := elem
					if l := len("refresh"); len(elem) >= l && elem[0:l] == "refresh" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "POST":
							r.name = RefreshTokenOperation
							r.summary = "Exchange a refresh token for a new access+refresh pair"
							r.operationID = "refresh-token"
							r.operationGroup = ""
							r.pathPattern = "/tokens/refresh"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

					elem = origElem
				case 'u': // Prefix: "user/"
					origElem := elem
					if l := len("user/"); len(elem) >= l && elem[0:l] == "user/" {
						elem = elem[l:]
					} else {
						break
					}

					// Param: "credential_id"
					// Leaf parameter, slashes are prohibited
					idx := strings.IndexByte(elem, '/')
					if idx >= 0 {
						break
					}
					args[0] = elem
					elem = ""

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "DELETE":
							r.name = RevokeUserTokensOperation
							r.summary = "Revoke all sessions for a credential without deleting it"
							r.operationID = "revoke-user-tokens"
							r.operationGroup = ""
							r.pathPattern = "/tokens/user/{credential_id}"
							r.args = args
							r.count = 1
							return r, true
						default:
							return
						}
					}

					elem = origElem
				case 'v': // Prefix: "validate"
					origElem := elem
					if l := len("validate"); len(elem) >= l && elem[0:l] == "validate" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "POST":
							r.name = ValidateTokenOperation
							r.summary = "Validate any token — JWT or OTU"
							r.operationID = "validate-token"
							r.operationGroup = ""
							r.pathPattern = "/tokens/validate"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

					elem = origElem
				}
				// Param: "jti"
				// Leaf parameter, slashes are prohibited
				idx := strings.IndexByte(elem, '/')
				if idx >= 0 {
					break
				}
				args[0] = elem
				elem = ""

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "DELETE":
						r.name = RevokeTokenOperation
						r.summary = "Revoke a single token by JTI (idempotent)"
						r.operationID = "revoke-token"
						r.operationGroup = ""
						r.pathPattern = "/tokens/{jti}"
						r.args = args
						r.count = 1
						return r, true
					default:
						return
					}
				}

			}

		}
	}
	return r, false
}